Enum nn::LearningMode

pub enum LearningMode {
    Incremental,
    Batch(u32),
}

Specifies which learning mode to use when training the network

Variants

Incremental

train the network Incrementally (updates weights after each example)

Batch

train the network in batch (updates weights only at then end of each epoch) batch training can be parallelized and thus the Batch constructor takes a u32 that specifies the number of threads to use when training the network