nn
Module norm

Specifies how to normalize and validate the result of a network.

Structs

ISigmoid

Normalize internal values using sigmoid.

OBitVec

Normalize the outputs with sigmoid. Error is computed as half of the sum of the squares of the individual errors.

OClassifier

Normalize outputs with softmax, and use cross-entropy to estimate the error. This is suitable for functions that need to pick one out of some options (i.e., classifiers). Each result represents the likelihood that the input belongs to the given class, and all outputs sum up to 1.

ORVec

Don’t normalize the outputs. Error is computed as half of the sum of the squares of the individual errors.

Traits

Internal

How to normalize the internal results computed by the net.

Norm

How to normalize values computed by the network.

Output

How to compute and validate the output of the net.