nn::norm
Trait Output

pub trait Output {
    // Required methods
    fn normalize(ys: &mut [f32]);
    fn error(actual: &[f32], expected: &[f32]) -> f32;
    fn error_delta(actual: &[f32], expected: &mut [f32]);
}

How to compute and validate the output of the net.

Required Methods

fn normalize(ys: &mut [f32])

Normalize the given results.

fn error(actual: &[f32], expected: &[f32]) -> f32

Compute the error for the given results.

fn error_delta(actual: &[f32], expected: &mut [f32])

Compute the error gradient. The error gradient is returned in the location where the expected values were provided.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.