Crate nn

Structs

Dim

Geometry of a neural net.

Learner

A neural net in training.

RunnerEmpty

Infrastructure for evaluating a net, without the weights.

RunnerReady

A neural net that can be used to map inputs to outputs.

Weights

The weights of a neural network.

Functions

sel

Square error loss function. Used during training to estimate how far off we are from a desired result.

sigmoid

Activation function for a neuron. Used to normalize each neuron’s output.

Type Aliases

R

The type of net weights, inputs, and outputs.