Crate nn

Structs

Dim

Geometry of a neural net.

Learner

A neural net in training.

OutputBitVec

Normalize the outputs with sigmoid, and use square error loss.

OutputClassifier

Normalize outputs with softmax, and use cross-entropy loss. This is suitable for functions that need to pick one out of some options (i.e., classifiers). Each result represents the likelihood that the input belongs to the given class, and all outputs sum up to 1.

OutputVec

Don’t normalize the outputs, and use square error loss.

Runner

A neural net that can be used to map inputs to outputs.

RunnerState

Infrastructure for evaluating a net, without the weights.

Weights

The weights of a neural network.

Traits

OutputNorm

How to compute and validate the output of the net.

Type Aliases

R

The type of net weights, inputs, and outputs.